Script for managing the asteroid pool
To save performance, asteroids were stored in a 'pool'. When they were needed they were made active and positioned where needed. When they were destroyed, they were infact just made in-active. This simple script is called when an asteroid of a certain size is requested, and at a target position if world co-ordinates are provided.